# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pengumuman API",
        "description": "API untuk mengelola pengumuman",
        "contact": {
            "name": "API Support",
            "email": "admin@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/pengumuman/terbaru": {
            "get": {
                "tags": [
                    "Pengumuman"
                ],
                "summary": "Mengambil 5 pengumuman terbaru",
                "operationId": "be9dbd23160bf0fd2bdbecb8dd3bf505",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data pengumuman terbaru",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data pengumuman terbaru berhasil diambil"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "judul": {
                                                        "type": "string",
                                                        "example": "Pembayaran UKT"
                                                    },
                                                    "isi": {
                                                        "type": "string",
                                                        "example": "Segera lakukan pembayaran UKT semester ganjil."
                                                    },
                                                    "tanggal": {
                                                        "type": "string",
                                                        "example": "26 Jun 2025"
                                                    },
                                                    "kategori": {
                                                        "type": "string",
                                                        "example": "keuangan"
                                                    },
                                                    "lokasi": {
                                                        "type": "string",
                                                        "example": "Bank Syariah Mandiri, Kampus Pusat"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Terjadi kesalahan server",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Terjadi kesalahan pada server"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/pengumuman": {
            "get": {
                "tags": [
                    "Pengumuman"
                ],
                "summary": "Mengambil semua pengumuman dengan pagination",
                "operationId": "20fe71a69229b75e4863b5c952a1f439",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Nomor halaman",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Jumlah data per halaman",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "kategori_id",
                        "in": "query",
                        "description": "Filter berdasarkan ID kategori",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data pengumuman",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data pengumuman berhasil diambil"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "judul": {
                                                        "type": "string",
                                                        "example": "Pembayaran UKT"
                                                    },
                                                    "isi": {
                                                        "type": "string",
                                                        "example": "Segera lakukan pembayaran UKT semester ganjil."
                                                    },
                                                    "tanggal": {
                                                        "type": "string",
                                                        "example": "26 Jun 2025"
                                                    },
                                                    "kategori": {
                                                        "type": "string",
                                                        "example": "keuangan"
                                                    },
                                                    "lokasi": {
                                                        "type": "string",
                                                        "example": "Bank Syariah Mandiri, Kampus Pusat"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "pagination": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 25
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 3
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/pengumuman/{id}": {
            "get": {
                "tags": [
                    "Pengumuman"
                ],
                "summary": "Mengambil detail pengumuman berdasarkan ID",
                "operationId": "240cd78f695233c0ed2d7c068f02295d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID pengumuman",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil detail pengumuman",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Detail pengumuman berhasil diambil"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "judul": {
                                                    "type": "string",
                                                    "example": "Pembayaran UKT"
                                                },
                                                "isi": {
                                                    "type": "string",
                                                    "example": "Segera lakukan pembayaran UKT semester ganjil."
                                                },
                                                "tanggal": {
                                                    "type": "string",
                                                    "example": "26 Jun 2025"
                                                },
                                                "kategori": {
                                                    "type": "string",
                                                    "example": "keuangan"
                                                },
                                                "lokasi": {
                                                    "type": "string",
                                                    "example": "Bank Syariah Mandiri, Kampus Pusat"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pengumuman tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Pengumuman tidak ditemukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}